fix: enable clipboard test cases in multiple editor
