+++ tests/multiple-editor/second-editor.spec.ts
@@ -1,5 +1,6 @@
 import './setup-second.js';
 import '../basic.case.ts?second';
+import '../clipboard.case.ts?second';
 
 import { setScope } from '../utils/playwright.js';
 setScope('');
